Extract logic to a new hook
