+++ packages/core/src/pages/_app.tsx
@@ -2,7 +2,6 @@ import { DefaultSeo } from 'next-seo'
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
-import { useEffect } from 'react'
 
 import { useSearch } from '@faststore/sdk'
 import { UIProvider } from '@faststore/ui'
@@ -13,8 +12,8 @@ import AnalyticsHandler from 'src/sdk/analytics'
 import { DeliveryPromiseProvider } from 'src/sdk/deliveryPromise'
 import ErrorBoundary from 'src/sdk/error/ErrorBoundary'
 import useGeolocation from 'src/sdk/geolocation/useGeolocation'
+import { useLocaleValidation } from 'src/sdk/i18n'
 import useScrollRestoration from 'src/sdk/ui/useScrollRestoration'
-import { validateLocaleForHostname } from 'src/utils/validateLocaleForHostname'
 
 import storeConfig from 'discovery.config'
 import SEO from 'next-seo.config'
@@ -34,29 +33,7 @@ function App({ Component, pageProps }: AppProps) {
   startGlobalSearchState(router.asPath, { itemsPerPage: ITEMS_PER_PAGE })
 
   // Client-side validation of locale binding (fallback for static pages)
-  useEffect(() => {
-    // Skip validation if we're already on the 404 page (from SSR notFound)
-    if (router.pathname === '/404') {
-      return
-    }
-
-    const hostname = window.location.hostname
-    const locale = router.locale
-
-    if (!locale) {
-      return
-    }
-
-    try {
-      const isValid = validateLocaleForHostname(hostname, locale)
-
-      if (!isValid) {
-        window.location.href = `/404?from=${encodeURIComponent(`/${locale}${router.asPath}`)}`
-      }
-    } catch (error) {
-      console.warn('Client-side locale validation failed:', error)
-    }
-  }, [router.locale, router.asPath, router.defaultLocale, router.pathname])
+  useLocaleValidation()
 
   return (
     <ErrorBoundary>
